+++ examples/eslint.config.js
@@ -1,4 +1,11 @@
 import { defineConfig } from "eslint/config";
 import baseConfig from "../eslint.config.js";
 
-export default defineConfig({ extends: [baseConfig] });
+export default defineConfig({
+  extends: [baseConfig],
+  rules: {
+    "@typescript-eslint/no-unused-vars": "off",
+    "@typescript-eslint/no-empty-function": "off",
+    "@typescript-eslint/no-unused-expressions": "off",
+  },
+});