+++ src/draft-assistant/index.ts
@@ -2,4 +2,3 @@ export * from "./DraftAssistant";
 export * from "./FieldDraftAssistant";
 export * from "./SectionDraftAssistant";
 export * from "./IntegerDraftAssistant";
-export * from "./DateDraftAssistant";