+++ src/rule/RulesBroken.ts
@@ -20,39 +20,36 @@ export interface RulesBrokenAsJson {
  * @category Rules
  */
 export class RulesBroken extends Error {
-  /**
-   * @category Construction
-   */
+  /** @category Creation */
   static fromJson(rulesBrokenAsJson: RulesBrokenAsJson) {
-    const brokenRules = rulesBrokenAsJson.brokenRules.map((ruleAsJson) => RuleLabel.fromJson(ruleAsJson));
+    const brokenRules = rulesBrokenAsJson.brokenRules.map((ruleAsJson) =>
+      RuleLabel.fromJson(ruleAsJson)
+    );
 
     return new this(brokenRules);
   }
 
-  /**
-   * @category Construction
-   */
+  /** @category Creation */
   constructor(protected brokenRules: LabeledRule[]) {
     super();
   }
 
-  /**
-   * @category Inspection
-   */
+  /** @category Inspection */
   hasRuleBrokenWith(labelId: LabelId, labelDescription: string) {
-    return this.brokenRules.some((rule) => rule.hasLabel(labelId, labelDescription));
+    return this.brokenRules.some((rule) =>
+      rule.hasLabel(labelId, labelDescription)
+    );
   }
 
-  /**
-   * @category Inspection
-   */
+  /** @category Inspection */
   hasOnlyOneRuleBrokenWith(labelId: LabelId, labelDescription: string) {
-    return this.brokenRules.length === 1 && this.brokenRules[0].hasLabel(labelId, labelDescription);
+    return (
+      this.brokenRules.length === 1 &&
+      this.brokenRules[0].hasLabel(labelId, labelDescription)
+    );
   }
 
-  /**
-   * @category Inspection
-   */
+  /** @category Inspection */
   forEachRuleBroken(closure: (brokenRule: LabeledRule) => void) {
     return this.brokenRules.forEach(closure);
   }