+++ src/rule/Ruleset.ts
@@ -11,34 +11,16 @@ import type {
  * Runs all rules and throws an error if any has failed.
  * The failed rules are included in the error.
  *
- * @see {@link RulesBroken}
- *
  * @category Rules
  */
 export class Ruleset {
   /**
    * Evaluates all assertions **synchronously** and throws an error if any has failed.
    *
-   * @throws if any rule has failed.
+   * @throws {@link RulesBroken} if any rule has failed.
    *
    * @example
-   * ```ts
-   * const firstNameNotEmptyAssertion = Assertion.requiring(
-   *    "firstName",
-   *    firstNameNotEmptyDescription,
-   *    Requirements.isNotEmpty
-   * );
-   * const lastNameNotEmptyAssertion  = Assertion.requiring(
-   *    "lastName",
-   *    lastNameNotEmptyDescription,
-   *    Requirements.isNotEmpty
-   * );
-   *
-   * Ruleset.ensureAll(
-   *   firstNameNotEmptyAssertion.evaluateFor("Jane"),
-   *   lastNameNotEmptyAssertion.evaluateFor(""),
-   * ); // throws because last name is empty
-   * ```
+   * {@includeCode ../../examples/snippets/rules.ts#ruleset-ensureAll}
    */
   static ensureAll(...assertions: SelfContainedAssertions[]): void {
     new this(assertions.flat(), []).ensure();
@@ -47,32 +29,19 @@ export class Ruleset {
   /**
    * Evaluates all rules **asynchronously** and throws an error if any has failed.
    *
-   * @throws if any rule has failed
+   * @throws {@link RulesBroken} if any rule has failed
    *
    * @example
-   * ```ts
-   * const emailMustBeUniqueInquiry = Inquiry.requiring(
-   *    "email",
-   *    emailMustBeUniqueDescription,
-   *    async (email) => !(await isEmailTaken(email))
-   * );
-   * const emailMustNotBeDisposableInquiry = Inquiry.requiring(
-   *    "email",
-   *    emailMustNotBeDisposableDescription,
-   *    async () => !(await isDisposable(email))
-   * );
-   *
-   * await Ruleset.workOn(
-   *   emailMustBeUniqueInquiry.evaluateFor("example@disposable.com"),
-   *   emailMustNotBeDisposableInquiry.evaluateFor("example@disposable.com"),
-   * ); // throws because email is disposable
-   * ```
+   * {@includeCode ../../examples/snippets/rules.ts#email-unique,ruleset-workOn}
    */
   static workOn(...rules: SelfContainedRules[]): Promise<void> {
     return new this([], rules.flat()).mustHold();
   }
 
-  constructor(protected assertions: SelfContainedAssertion[], protected rules: SelfContainedRule[]) {}
+  constructor(
+    protected assertions: SelfContainedAssertion[],
+    protected rules: SelfContainedRule[]
+  ) {}
 
   async mustHold(): Promise<void> {
     const brokenRules = await this.brokenRules();
@@ -88,7 +57,9 @@ export class Ruleset {
 
   protected failedAssertions(): LabeledRule[] {
     const failed: LabeledRule[] = [];
-    this.assertions.forEach((assertion) => assertion.collectFailureInto(failed));
+    this.assertions.forEach((assertion) =>
+      assertion.collectFailureInto(failed)
+    );
     return failed;
   }