+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
+## [1.0.0] - 2025-05-03
+
+### Removed
+
+- `DateDraftAssistant` class.
+- `SectionDraftAssistant.routeFailedAssertionsOf`
+- `SectionDraftAssistant.routeFailedAssertion`
+
 ## [0.7.2] - 2025-05-01
 
 ### Added